requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-ninja==1.1.0
idna==3.7
pydantic==2.7.4
pydantic_core==2.18.4
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
